refactor TensorDataCollatorWithFlattening
